[IMP] account_financial_report: abstract for avoiding duplicated code and increase readability
