+++ account_financial_report/report/__init__.py
@@ -3,6 +3,7 @@
 # © 2016 Julien Coux (Camptocamp)
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).-
 
+from . import abstract_report
 from . import abstract_report_xlsx
 from . import aged_partner_balance
 from . import aged_partner_balance_xlsx
