+++ account_financial_report/wizard/abstract_wizard.py
@@ -34,3 +34,18 @@ class AbstractWizard(models.AbstractModel):
         required=False,
         string="Company",
     )
+
+    def button_export_html(self):
+        self.ensure_one()
+        report_type = "qweb-html"
+        return self._export(report_type)
+
+    def button_export_pdf(self):
+        self.ensure_one()
+        report_type = "qweb-pdf"
+        return self._export(report_type)
+
+    def button_export_xlsx(self):
+        self.ensure_one()
+        report_type = "xlsx"
+        return self._export(report_type)
