+++ account_financial_report/wizard/journal_ledger_wizard.py
@@ -94,20 +94,6 @@ class JournalLedgerReportWizard(models.TransientModel):
             .report_action(self, data=data)
         )
 
-    def button_export_html(self):
-        self.ensure_one()
-        report_type = "qweb-html"
-        return self._export(report_type)
-
-    def button_export_pdf(self):
-        report_type = "qweb-pdf"
-        return self._export(report_type)
-
-    def button_export_xlsx(self):
-        self.ensure_one()
-        report_type = "xlsx"
-        return self._export(report_type)
-
     def _prepare_report_journal_ledger(self):
         self.ensure_one()
         journals = self.journal_ids
