+++ account_financial_report/wizard/vat_report_wizard.py
@@ -8,13 +8,8 @@ from odoo.exceptions import ValidationError
 class VATReportWizard(models.TransientModel):
     _name = "vat.report.wizard"
     _description = "VAT Report Wizard"
+    _inherit = "account_financial_report_abstract_wizard"
 
-    company_id = fields.Many2one(
-        comodel_name="res.company",
-        default=lambda self: self.env.company.id,
-        required=False,
-        string="Company",
-    )
     date_range_id = fields.Many2one(comodel_name="date.range", string="Date range")
     date_from = fields.Date("Start Date", required=True)
     date_to = fields.Date("End Date", required=True)
@@ -87,21 +82,6 @@ class VATReportWizard(models.TransientModel):
             )
             .report_action(self, data=data)
         )
-
-    def button_export_html(self):
-        self.ensure_one()
-        report_type = "qweb-html"
-        return self._export(report_type)
-
-    def button_export_pdf(self):
-        self.ensure_one()
-        report_type = "qweb-pdf"
-        return self._export(report_type)
-
-    def button_export_xlsx(self):
-        self.ensure_one()
-        report_type = "xlsx"
-        return self._export(report_type)
 
     def _prepare_vat_report(self):
         self.ensure_one()
